requirements.txt
torch
torchvision
torchaudio
pandas
numpy
matplotlib
scikit-learn
pillow
tqdm
opencv-python
